Scroll to active item when list is expanded
List expansion only happens when a new item is added and as such the `onHeightChanged` signal allows for bringing the new item into view.
This work around is necessary as all my attempts to scroll with the output by e.g. using the `onExecuted` signal of `TerminalItem` failed.
Reasons for this include both the async nature of JavaScript which makes it inconvenient to enforce execution of a function at exactly the right time, as well as the height value of `terminalList` seemingly not being updated instantly.