Change color setting types to `string`
While `color` would obviously be the correct type to store a color value it is serialized as a `QVariant` value which is not easily changeable by the user.
This requirement is satisfied by `string` and as such this type is - at this point in time - a superior replacement of `color`.
Should the QML settings module be changed to serialize color values in a human readable fashion these types should be changed back to their formally correct value.