Reimplemented `kill` command using actual signal sending
Previously `kill` merely destroyed the terminal widget. This is contrary to the expected notion of `kill`, i.e. actually killing the process explicitly instead of relying on side effects.

The normal mode `d` verb now explicitly destroys the current item's content, be it a terminal widget or a history view.