Add initial working directory to settings handler
In addition to increasing customizability this change makes it clear that the _working directory_ property only concerns the intial working directory and is not updated if the contained process changes its directory.

Note that setting the intial working directory is broken in upstream _QMLTermWidget_ and the appropriate patch [1] is not merged at the time of this commit.

[1]: https://github.com/KnairdA/qmltermwidget/tree/initial_working_directory_fix